feat(thresh): add WeightedThreshold to KeyConfig
