+++ Cargo.toml
@@ -40,6 +40,7 @@ arrayref = "0.3.6"
 zeroize = "1.3.0"
 sled = { version = "0.34.6", optional = true }
 fixed = { version = "1.9", optional = true }
+fraction = { version = "0.8"}
 
 rkv = { version = "0.17", optional = true }
 bincode = { version = "1.3.1", optional = true }
